+++ package.json
@@ -20,6 +20,7 @@
     "react-router-dom": "^5.2.0",
     "react-saga": "^0.3.1",
     "react-scripts": "^5.0.1",
+    "react-scroll-horizontal": "^1.6.6",
     "redux": "^4.0.5",
     "redux-actions": "^2.6.5",
     "redux-devtools-extension": "^2.13.8",
